feat: Add support for 'rust_version'
This was added to the index in rust-lang/crates.io#6267.  crates.io is
automatically injecting it into the index, even without using the
nightly cargo that supports it.

My plan is to use this to make cargo-upgrade more MSRV aware (in version
reqs despite the resolver not yet being MSRV aware).