fix: remove the usage of file locks in preference for documentation when opening a git index.
Previously, to allow concurrently opening and possibly updating a crates-index, a file-lock was
used for synchronization. However, it was rather specific to what the test-suite needed while
adding another failure mode for production code which could leave lock-files behind that then
lock the crates-index forever for this library at least.

Instead, appropriate locking will be used in tests only, while the documentation of all
`open` methods of `GitIndex` was adjusted to inform about ways to protect concurrent accesses
on application level.