fix: assure Git index updates to refs are actually written.
The remote git repository may alter its references in such a way that
local fast-forwards aren't possible anymore.

This happens regularly as the history will be squashed on the remote.

Now we forcefully store the updated references, which resolves
the issue that calling `update()` didn't seem to do anything despite
being busy (i.e. downloading a possibly huge pack, and resolving it).